+++ src/lib.rs
@@ -231,5 +231,6 @@ pub use shred::AsyncDispatcher;
 
 pub use changeset::ChangeSet;
 pub use storage::{DenseVecStorage, FlaggedStorage, HashMapStorage, InsertedFlag, ModifiedFlag,
-                  NullStorage, ReadStorage, RemovedFlag, Storage, Tracked, VecStorage, WriteStorage};
+                  NullStorage, ReadStorage, RemovedFlag, Storage, Tracked, VecStorage,
+                  WriteStorage};
 pub use world::{Builder, Component, Entities, Entity, EntityBuilder, LazyUpdate, World};