+++ src/saveload/marker.rs
@@ -79,8 +79,8 @@ impl<'a> MarkedBuilder for LazyBuilder<'a> {
     /// Panics during [`world.maintain()`](World::maintain) in case there's no allocator
     /// added to the [`World`].
     fn marked<M>(self) -> Self
-        where
-            M: Marker
+    where
+        M: Marker
     {
         let entity = self.entity;
         self.lazy.exec(move |world| {
@@ -222,8 +222,7 @@ impl<'a> EntityResBuilder<'a> {
 ///     );
 /// }
 /// ```
-pub trait Marker
-    : Clone + Component + Debug + Eq + Hash + DeserializeOwned + Serialize {
+pub trait Marker: Clone + Component + Debug + Eq + Hash + DeserializeOwned + Serialize {
     /// Id of the marker
     type Identifier;
     /// Allocator for this `Marker`