+++ CHANGELOG.md
@@ -1,5 +1,8 @@
-# 0.12.0
+# [Unreleased]
 
+* `Join::open()` and `Storage::unprotected_storage_mut()` have been marked unsafe.
+Thanks to [@andrewhickman](https://github.com/andrewhickman) for discovering this
+unsoundness!
 * Add common `Builder` trait to `EntityBuilder` and `LazyBuilder` ([#426])
 * Add common `MarkedBuilder` trait to `EntityBuilder` and `LazyBuilder` ([#426])